introduction-to-jaeger-tracing/requirements.txt
opentelemetry-api
opentelemetry-exporter-jaeger
